Telephony: Add subaddress to SMS fingerprint calculation.
As per SMS specification in 3GPP2 C.S0015-B, section 4.3.1.6, if the
Subaddress is included in a CDMA SMS message, it needs to be used for
duplicate detection. Subaddress, which is an optional field was omitted
while computing the SMS fingerprint. Hence it was never being used in
duplicate detection if it was included in the SMS. Add subaddress to the
SMS fingerprint.

Change-Id: I1b69dfb16aef86e5587a467ec681af7692e27e6a
CRs-Fixed: 235885