STK: Create STK service with context and command parameters alone.
This change starts the STK service with minimal parameters (context
and Command Interface). The rest of the parameters are
updated later when they are available so that the applications (STK app)
which require to get an instance of the STK service need not wait
until ICCFilehandler is ready (this may only be ready after
substantial delay).

Change-Id: I18b47c78db45f6fd5b54cabc3ab783649ff7f114